Add normalizeSvgIconName utility
